+++ Client/Assets/Scripts/Data/MetaData.cs
@@ -37,8 +37,8 @@ public class EquipmentMetaData
     public int Id;
     public EquipmentType EquipmentType;
     public EquipSlotType EquipSlotType;
-    public float Attack;
-    public float Defense;
+    public int Attack;
+    public int Defense;
     public int RequiredLevel;
 }
 
@@ -102,9 +102,9 @@ public class ProjectileMetaData
     public int LifeTime;
     public int MaxHitCount;
     public float MoveSpeed;
-    public float CommonAttackDamage;
+    public int CommonAttackDamage;
     public float CommonAttackCoolTime;
-    public float MaxHp;
+    public int MaxHp;
 }
 
 // ── Exp
@@ -124,10 +124,10 @@ public class MonsterMetaData
     public MonsterType MonsterType;
     public int DropItemGroupId;
     public int MaxHp;
-    public float CommonAttackDamage;
+    public int CommonAttackDamage;
     public float CommonAttackCoolTime;
     public float AttackRange;
-    public float Defense;
+    public int Defense;
     public float MoveSpeed;
     public int Level;
     public int Exp;
@@ -155,10 +155,10 @@ public class PlayerMetaData
     public int Id;
     public PlayerJobType PlayerJobType;
     public int MaxHp;
-    public float CommonAttackDamage;
+    public int CommonAttackDamage;
     public float CommonAttackCoolTime;
     public float AttackRange;
-    public float Defense;
+    public int Defense;
     public float MoveSpeed;
     public int AttackHalfAngleDeg;
     public int AttackHeight;
